Energy history vector issue is fixed.
The problem was with memory management. I tried to update and int with a
function, which obviously failed because of scope issues. I realised
however that I could simply pass a pointer to the structure and then use
the -> command to access the structure's attributes. Problem solved and
a new technique learned!